modified-房間預訂管理，調整新增編輯房間預定日期不能重疊在一起

+++ Application/ViewModels/RoomType/RoomTypeViewModel.cs
@@ -25,11 +25,11 @@ namespace Application.ViewModels.RoomType
 
         [Required(ErrorMessage = "{0}必填")]
         [Display(Name = "平日價")]
-        public decimal? Price { get; set; }
+        public int? Price { get; set; }
 
         [Required(ErrorMessage = "{0}必填")]
         [Display(Name = "假日價")]
-        public decimal? Hprice { get; set; }
+        public int? Hprice { get; set; }
 
         [Required]
         [Display(Name = "是否有空調")]
